requirements.txt
# Binance SDK
python-binance

# 
requests

# Compatibility consideration (local env)
python-dotenv